requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
alembic
pydantic[email]
python-dotenv
